refactor(encoding): extract shared match-finder helpers into match_table
Move the four standalone helpers used by every match finder
(`best_len_offset_candidate`, `extend_backwards_shared`,
`repcode_candidate_shared`, `pick_lazy_match_shared`) plus the
`LazyMatchConfig` knobs they take out of `match_generator.rs` into
`match_table::helpers`. Mechanical move only — names, signatures, and
bodies are byte-for-byte identical; the only intentional API change is
`pub(crate)` visibility on the relocated items so subsequent
per-matcher extractions (Simple / Dfast / Row / Hc into their own
modules) can import them through `super::match_table::helpers::*`
without re-introducing the helpers per matcher.

`MatchGenerator::common_prefix_len` is now `pub(crate)` so the
relocated `repcode_candidate_shared` can keep calling into the
Simple-matcher fastpath wrapper without taking a behaviour-changing
detour through `fastpath::dispatch_common_prefix_len_ptr` directly.

`match_generator.rs` shrinks by 168 lines (10080 → ~9905) and gains
the import block at the top. Build, clippy, and the level22 ratio
gate stay green.

Part of #111 Phase 1b.
